added admin module and shipping info component. fixed add to cart and retrieve cart methods

+++ frontend/glassify-frontend/src/app/cart/cart.service.ts
@@ -1,6 +1,7 @@
 import { Injectable } from '@angular/core';
 import { HttpClient } from '@angular/common/http';
-import { Observable } from 'rxjs';
+import { Observable,  of} from 'rxjs';
+import { switchMap } from 'rxjs/operators';
 import { tap } from 'rxjs/operators';
 import { environment } from 'src/environments/environment';
 import { CartItem } from '../model/cart-item.model';
@@ -9,42 +10,59 @@ import { CartItem } from '../model/cart-item.model';
   providedIn: 'root'
 })
 export class CartService {
+
   private apiUrl = environment.springUrl;
-  private cartId: string | null = null;
+  private cartIdKey = 'cartId';
 
   constructor(private http: HttpClient) { }
 
-  createCart(): Observable<string> {
-    return this.http.post<string>(`${this.apiUrl}/cart/create`, {});
-  }
-
-  addItemToCart(cartId: string, item: CartItem): Observable<CartItem> {
-    return this.http.post<CartItem>(`${this.apiUrl}/cart/${cartId}/add`, item);
-  }
-
-  removeItemFromCart(cartId: string, itemId: string): Observable<void> {
-    return this.http.delete<void>(`${this.apiUrl}/cart/${cartId}/remove/${itemId}`);
-  }
-
-  clearCart(cartId: string): Observable<void> {
-    return this.http.delete<void>(`${this.apiUrl}/cart/${cartId}/clear`);
-  }
-
-  getCartItems(cartId: string): Observable<CartItem[]> {
-    return this.http.get<CartItem[]>(`${this.apiUrl}/cart/${cartId}/getAll`);
-  }
-
   getCartId(): string | null {
-    return this.cartId;
+    return sessionStorage.getItem(this.cartIdKey);
   }
 
-  setCartId(id: string): void {
-    this.cartId = id;
-  }
-
-  createAndStoreCart(): Observable<string> {
-    return this.createCart().pipe(
-      tap(cartId => this.setCartId(cartId))
+  createCart(): Observable<string> {
+    return this.http.post<string>(`${this.apiUrl}/cart/create`, {}).pipe(
+      switchMap(cartId => {
+        sessionStorage.setItem(this.cartIdKey, cartId);
+        return of(cartId);
+      })
     );
   }
+
+  addItemToCart(item: CartItem): Observable<CartItem> {
+    let cartId = this.getCartId();
+    if (!cartId) {
+      return this.createCart().pipe(
+        switchMap(newCartId => {
+          return this.http.post<CartItem>(`${this.apiUrl}/cart/${newCartId}/add`, item);
+        })
+      );
+    } else {
+      return this.http.post<CartItem>(`${this.apiUrl}/cart/${cartId}/add`, item);
+    }
+  }
+
+  removeItemFromCart(itemId: string): Observable<void> {
+    let cartId = this.getCartId();
+    if (cartId) {
+      return this.http.delete<void>(`${this.apiUrl}/cart/${cartId}/items/${itemId}`);
+    } else {
+      return of(undefined);
+    }
+  }
+
+  clearCart(): Observable<void> {
+    let cartId = this.getCartId();
+    if (cartId) {
+    return this.http.delete<void>(`${this.apiUrl}/cart/${cartId}/clear`);
+    } else throw new Error("no cart created");
+  }
+
+  getCartItems(): Observable<CartItem[]> {
+    let cartId = this.getCartId();
+    if (cartId) {
+      return this.http.get<CartItem[]>(`${this.apiUrl}/cart/${cartId}/getAll`);
+    } else throw new Error("no cart created");
+  }
+
 }
